msp430: twi: move interrupt handler for USCI_Bx to twi.c
For the USCI module, it shares interrupt vector for
USCI_Ax and USCI_Bx, so the interrupt routine should still
be in usci_isr_handler.

But for USCI_B0 and USCI_B1, the interrupt handler can be
moved to twi.c. This is exactly the same way it is done
for the EUSCI_Bx module, it already has the interrupt
handler in twi.c